avec nos modules

+++ jpeg/include/dct_zig_zag_quantification.h
@@ -6,8 +6,8 @@
 
 extern float** precalculcos(int32_t largeur, int32_t longueur):
 
-extern int32_t* operations_dct_quantification_puis_zig_zag(int32_t** data, int cc, float** cosinus);
-extern int32_t* operations_naives(int32_t** data, int cc);
+extern int32_t* operations_dct_quantification_puis_zig_zag(uint8_t** data, int cc, float** cosinus);
+extern int32_t* operations_naives(uint8_t** data, int cc);
 
 
 #endif
